Fixes around feature-flagging, particularly in oxql, as well as uuid-kinds, sled-storage
