+++ nexus/Cargo.toml
@@ -51,7 +51,7 @@ num-integer.workspace = true
 once_cell.workspace = true
 openssl.workspace = true
 oximeter-client.workspace = true
-oximeter-db.workspace = true
+oximeter-db = { workspace = true, default-features = false, features = [ "oxql" ] }
 oxnet.workspace = true
 parse-display.workspace = true
 paste.workspace = true
